fix: solana indexer logic
