+++ omni-relayer/src/utils/mod.rs
@@ -1,4 +1,5 @@
 pub mod evm;
 pub mod near;
 pub mod redis;
+pub mod solana;
 pub mod storage;
